Introduced database seeder
This will use Bogus to create test raffle and participant data. The seeder bypass the `EntryService` since we want it to be able to manipulate the date and time.
The seeder is only available in development. The `acme.http` file is updated with a query that do some seeding.
There is also a change to App.razor since we needed CascadingAuthenticationState to wrap all Razor pages in order to POST.